+++ Model/Source/AbstractTable.php
@@ -105,6 +105,10 @@ abstract class AbstractTable extends AbstractFieldArray
             }
         }
 
+        if ($row['_id'] === null || is_int($row['_id'])) {
+            $row->setData('_id', '_' . random_int(1000000000, 9999999999) . '_' . random_int(0, 999));
+        }
+
         $row->setData('option_extra_attrs', $options);
     }
 }